requirements.txt
#EbookLib==0.18
#git+https://github.com/changyy/ebooklib.git@release#egg=EbookLib
lxml==4.9.3
Pillow==10.1.0
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
six==1.16.0
